Fix printf warnings
- Use length specifier "z" for size_t type and "t" for ptrdiff_t.

+++ SurgSim/Testing/MlcpIO/ReadText.cpp
@@ -267,8 +267,8 @@ static bool readEigenMatrix(const std::string& fileName, FILE* in, const char* l
 			ptrdiff_t newNumRows = matrix->rows() + 1;
 			if (rowVector.cols() != numCols)
 			{
-				fprintf(stderr, "Inconsistent number of columns for Eigen matrix (%d vs %d)\n  in file '%s'\n",
-						numCols, static_cast<int>(rowVector.cols()), fileName.c_str());
+				fprintf(stderr, "Inconsistent number of columns for Eigen matrix (%td vs %td)\n  in file '%s'\n",
+						numCols, rowVector.cols(), fileName.c_str());
 				return false;
 			}
 			matrix->conservativeResize(newNumRows, numCols);
